feat: ethereum-steth-v2 moved mocks
