+++ ethereum-steth-v2/package.json
@@ -55,11 +55,11 @@
     "fp-ts": "^2.16.1",
     "lodash": "^4.17.21",
     "ts-retry": "^4.2.4",
-    "winston": "^3.11.0"
+    "winston": "^3.11.0",
+    "@ethersproject/abi": "^5.0.0",
+    "@ethersproject/providers": "^5.0.0"
   },
   "devDependencies": {
-    "@ethersproject/abi": "^5.0.0",
-    "@ethersproject/providers": "^5.0.0",
     "@faker-js/faker": "^8.3.1",
     "@jest/globals": "^29.7.0",
     "@tsconfig/node20": "^20.1.2",
