+++ ethereum-steth-v2/src/utils/events/burner_events.ts
@@ -1,9 +1,7 @@
 import { EventOfNotice } from '../../entity/events'
-import { FindingSeverity, FindingType, LogDescription } from 'forta-agent'
+import { FindingSeverity, FindingType } from 'forta-agent'
 import { Result } from '@ethersproject/abi/lib'
 import { etherscanAddress } from '../string'
-import { faker } from '@faker-js/faker'
-import { createLogDescriptionMock } from '../contract_mocks/log_description.mock'
 
 export function getBurnerEvents(BURNER_ADDRESS: string): EventOfNotice[] {
   return [
@@ -34,28 +32,4 @@ export function getBurnerEvents(BURNER_ADDRESS: string): EventOfNotice[] {
       type: FindingType.Info,
     },
   ]
-}
-
-export function getFilteredBurnerEventsMock(): LogDescription[] {
-  const descriptions = [
-    {
-      ['requestedBy']: faker.finance.ethereumAddress(),
-      ['token']: faker.finance.ethereumAddress(),
-      ['amount']: faker.number.int(),
-    },
-    {
-      ['requestedBy']: faker.finance.ethereumAddress(),
-      ['token']: faker.finance.ethereumAddress(),
-      ['tokenId']: faker.number.int(),
-    },
-  ]
-
-  const out: LogDescription[] = []
-  for (const desc of descriptions) {
-    // eslint-disable-next-line
-    // @ts-expect-error
-    out.push(createLogDescriptionMock(desc))
-  }
-
-  return out
 }
