+++ ethereum-steth-v2/src/utils/events/lido_events.ts
@@ -1,9 +1,7 @@
-import { FindingSeverity, FindingType, LogDescription } from 'forta-agent'
+import { FindingSeverity, FindingType } from 'forta-agent'
 import { EventOfNotice } from '../../entity/events'
 import { Result } from '@ethersproject/abi/lib'
 import { etherscanAddress } from '../string'
-import { faker } from '@faker-js/faker'
-import { createLogDescriptionMock } from '../contract_mocks/log_description.mock'
 
 export const alertId_token_rebased = 'LIDO-TOKEN-REBASED'
 
@@ -108,39 +106,4 @@ export function getLidoEvents(LIDO_STETH_ADDRESS: string): EventOfNotice[] {
       type: FindingType.Info,
     },
   ]
-}
-export function getFilteredLidoEventsMock(): LogDescription[] {
-  const descriptions = [
-    {},
-    {},
-    {},
-    {},
-    {
-      ['maxStakeLimit']: faker.number.int(),
-      ['stakeLimitIncreasePerBlock']: faker.number.int(),
-    },
-    {},
-    {
-      ['lidoLocator']: faker.finance.ethereumAddress(),
-    },
-    {
-      ['vault']: faker.finance.ethereumAddress(),
-      ['token']: faker.finance.ethereumAddress(),
-    },
-    {
-      ['version']: faker.system.semver(),
-    },
-    {
-      ['amount']: faker.number.int(),
-    },
-  ]
-
-  const out: LogDescription[] = []
-  for (const desc of descriptions) {
-    // eslint-disable-next-line
-    // @ts-expect-error
-    out.push(createLogDescriptionMock(desc))
-  }
-
-  return out
 }
