+++ ethereum-steth-v2/src/utils/events/withdrawals_events.ts
@@ -1,9 +1,7 @@
-import { FindingSeverity, FindingType, LogDescription } from 'forta-agent'
+import { FindingSeverity, FindingType } from 'forta-agent'
 import BigNumber from 'bignumber.js'
 import { Result } from '@ethersproject/abi/lib'
 import { EventOfNotice } from '../../entity/events'
-import { faker } from '@faker-js/faker'
-import { createLogDescriptionMock } from '../contract_mocks/log_description.mock'
 
 export const WITHDRAWALS_BUNKER_MODE_ENABLED_EVENT = 'event BunkerModeEnabled(uint256 _sinceTimestamp)'
 
@@ -42,22 +40,4 @@ export function getWithdrawalsEvents(WITHDRAWAL_QUEUE_ADDRESS: string): EventOfN
       type: FindingType.Info,
     },
   ]
-}
-
-export function getFilteredWithdrawalsEventsMock(): LogDescription[] {
-  const descriptions = [
-    {},
-    {
-      ['duration']: faker.number.int(),
-    },
-  ]
-
-  const out: LogDescription[] = []
-  for (const desc of descriptions) {
-    // eslint-disable-next-line
-    // @ts-expect-error
-    out.push(createLogDescriptionMock(desc))
-  }
-
-  return out
 }
